+++ cdl/core/computation/signal.py
@@ -18,6 +18,7 @@ from __future__ import annotations
 import guidata.dataset as gds
 import numpy as np
 import scipy.integrate as spt
+import scipy.interpolate
 import scipy.ndimage as spi
 import scipy.optimize as spo
 import scipy.signal as sps
@@ -586,3 +587,76 @@ def compute_fw1e2(signal: SignalObj):
         yhm = amplitude / np.e**2 + base
         res.append([i_roi, mu - hw, yhm, mu + hw, yhm])
     return np.array(res)
+
+
+class InterpolationParam(gds.DataSet):
+    """Interpolation parameters"""
+
+    _methods = (
+        ("linear", _("Linear")),
+        ("spline", _("Spline")),
+        ("quadratic", _("Quadratic")),
+        ("cubic", _("Cubic")),
+        ("barycentric", _("Barycentric")),
+        ("pchip", _("PCHIP")),
+    )
+    method = gds.ChoiceItem(_("Interpolation method"), _methods, default="linear")
+    fill_value = gds.FloatItem(
+        _("Fill value"),
+        default=None,
+        help=_(
+            "Value to use for points outside the interpolation domain (used only "
+            "with linear, cubic and pchip methods)."
+        ),
+        check=False,
+    )
+
+
+def compute_interpolation(
+    src1: SignalObj, src2: SignalObj, p: InterpolationParam
+) -> SignalObj:
+    """Interpolate data
+    Args:
+        src1 (SignalObj): source signal 1
+        src2 (SignalObj): source signal 2
+        p (InterpolationParam): parameters
+    Returns:
+        SignalObj: result signal object
+    """
+    suffix = f"method={p.method}"
+    if p.fill_value is not None:
+        suffix += f", fill_value={p.fill_value}"
+    dst = dst_n1n(src1, src2, "interpolation", suffix)
+    x1, y1 = src1.get_data()
+    xnew, _y2 = src2.get_data()
+    interpolator_extrap = None
+    if p.method == "linear":
+        # Linear interpolation using NumPy's interp function:
+        ynew = np.interp(xnew, x1, y1, left=p.fill_value, right=p.fill_value)
+    elif p.method == "spline":
+        # Spline using 1-D interpolation with SciPy's interpolate package:
+        knots, coeffs, degree = scipy.interpolate.splrep(x1, y1, s=0)
+        ynew = scipy.interpolate.splev(xnew, (knots, coeffs, degree), der=0)
+    elif p.method == "quadratic":
+        # Quadratic interpolation using NumPy's polyval function:
+        coeffs = np.polyfit(x1, y1, 2)
+        ynew = np.polyval(coeffs, xnew)
+    elif p.method == "cubic":
+        # Cubic interpolation using SciPy's Akima1DInterpolator class:
+        interpolator_extrap = scipy.interpolate.Akima1DInterpolator(x1, y1)
+    elif p.method == "barycentric":
+        # Barycentric interpolation using SciPy's BarycentricInterpolator class:
+        interpolator = scipy.interpolate.BarycentricInterpolator(x1, y1)
+        ynew = interpolator(xnew)
+    elif p.method == "pchip":
+        # PCHIP interpolation using SciPy's PchipInterpolator class:
+        interpolator_extrap = scipy.interpolate.PchipInterpolator(x1, y1)
+    else:
+        raise ValueError(f"Invalid interpolation method {p.method}")
+    if interpolator_extrap is not None:
+        ynew = interpolator_extrap(xnew, extrapolate=p.fill_value is None)
+        if p.fill_value is not None:
+            ynew[xnew < x1[0]] = p.fill_value
+            ynew[xnew > x1[-1]] = p.fill_value
+    dst.set_xydata(xnew, ynew)
+    return dst